% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzdb-package.R
\docType{package}
\name{tzdb-package}
\alias{tzdb}
\alias{tzdb-package}
\title{tzdb: Time Zone Database Information}
\description{
Provides an up-to-date copy of the Internet Assigned Numbers
    Authority (IANA) Time Zone Database. It is updated periodically to
    reflect changes made by political bodies to time zone boundaries, UTC
    offsets, and daylight saving time rules. Additionally, this package
    provides C++ headers and callables for working with the 'date'
    library. 'date' provides comprehensive support for working with dates
    and date-times, which this package exposes to make it easier for other
    R packages to utilize. Headers are provided for calendar specific
    calculations, along with a limited interface for time zone
    manipulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/tzdb}
  \item Report bugs at \url{https://github.com/r-lib/tzdb/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item Howard Hinnant (Author of the included date library) [copyright holder]
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
