% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{tzdb_names}
\alias{tzdb_names}
\title{Time zone database names}
\usage{
tzdb_names()
}
\value{
A character vector of zone names.
}
\description{
\code{tzdb_names()} returns the time zone names found in the database.
}
\examples{
tzdb_names()
}
