% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{tzdb_path}
\alias{tzdb_path}
\title{Retrieve the path to the time zone database}
\usage{
tzdb_path(type)
}
\arguments{
\item{type}{\verb{[character(1)]}

The type of time zone database to return the path for.

Currently only \code{"text"} is supported.}
}
\value{
A single string containing the path to the database.
}
\description{
Retrieve the path to the time zone database
}
\examples{
tzdb_path("text")
}
