% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{tzdb_version}
\alias{tzdb_version}
\title{Retrieve the time zone database version}
\usage{
tzdb_version()
}
\value{
A single string containing the version number.
}
\description{
Retrieve the time zone database version
}
\examples{
tzdb_version()
}
