% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{tzdb_version}
\alias{tzdb_version}
\title{Time zone database version}
\usage{
tzdb_version()
}
\value{
A single string of the database version.
}
\description{
\code{tzdb_version()} returns the version of the time zone database currently in
use.
}
\examples{
tzdb_version()
}
